Add new `extension` crate
