+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "crates/docs_preprocessor",
     "crates/editor",
     "crates/evals",
+    "crates/extension",
     "crates/extension_api",
     "crates/extension_cli",
     "crates/extension_host",
@@ -201,6 +202,7 @@ copilot = { path = "crates/copilot" }
 db = { path = "crates/db" }
 diagnostics = { path = "crates/diagnostics" }
 editor = { path = "crates/editor" }
+extension = { path = "crates/extension" }
 extension_host = { path = "crates/extension_host" }
 extensions_ui = { path = "crates/extensions_ui" }
 feature_flags = { path = "crates/feature_flags" }
