+++ crates/extension/src/extension_manifest.rs
